requirements.txt
duckdb==1.1.3
pandas==2.2.3
plotly==6.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
streamlit==1.41.1
pytest==8.3.4
types-psycopg2
databricks-sql-connector>=3.0.0